+++ cmd/forklift/stage/bundles.go
@@ -13,7 +13,7 @@ import (
 // ls-bundle
 
 func lsBunAction(c *cli.Context) error {
-	store, err := getStageStore(c.String("workspace"), false)
+	store, err := getStageStore(c.String("workspace"))
 	if err != nil {
 		return err
 	}
@@ -48,7 +48,7 @@ func lsBunAction(c *cli.Context) error {
 // show-bun
 
 func showBunAction(c *cli.Context) error {
-	store, err := getStageStore(c.String("workspace"), false)
+	store, err := getStageStore(c.String("workspace"))
 	if err != nil {
 		return err
 	}
@@ -67,32 +67,4 @@ func showBunAction(c *cli.Context) error {
 	}
 	fcli.PrintStagedBundle(0, store, bundle, index)
 	return nil
-}
-
-// show-bun-depl
-
-func showBunDeplAction(c *cli.Context) error {
-	store, err := getStageStore(c.String("workspace"), false)
-	if err != nil {
-		return err
-	}
-	if !store.Exists() {
-		return errMissingStore
-	}
-
-	rawIndex := c.Args().First()
-	index, err := strconv.Atoi(rawIndex)
-	if err != nil {
-		return errors.Wrapf(err, "Couldn't parse staged bundle index %s as an integer", rawIndex)
-	}
-	deplName := c.Args().Get(1)
-	bundle, err := store.LoadFSBundle(index)
-	if err != nil {
-		return errors.Wrapf(err, "couldn't load staged bundle %d", index)
-	}
-	resolved, err := bundle.LoadDepl(deplName)
-	if err != nil {
-		return errors.Wrapf(err, "couldn't load deployment %s from bundle %d", deplName, index)
-	}
-	return fcli.PrintResolvedDepl(0, bundle, resolved)
 }